Extract class refactoring: Extracting email generator to a different class from internet
