+++ src/main/java/net/datafaker/providers/base/BaseFaker.java
@@ -423,4 +423,6 @@ public class BaseFaker implements BaseProviders {
         if (ap == null || (map = METHODS.get(ap.getClass())) == null) return null;
         return map.get(methodName);
     }
+
+
 }
